--- a/pkg/streamingpromql/operators/subquery.go
+++ b/pkg/streamingpromql/operators/subquery.go
@@ -25,9 +25,8 @@ type Subquery struct {
 	nextStepT         int64
 	rangeMilliseconds int64
 	floats            *types.FPointRingBuffer
-	floatView         *types.FPointRingBufferView
 	histograms        *types.HPointRingBuffer
-	histogramView     *types.HPointRingBufferView
+	stepData          *types.RangeVectorStepData // Retain the last step data instance we used to avoid allocating it for every step.
 }
 
 var _ types.RangeVectorOperator = &Subquery{}
@@ -51,6 +50,7 @@ func NewSubquery(
 		rangeMilliseconds:    subqueryRange.Milliseconds(),
 		floats:               types.NewFPointRingBuffer(memoryConsumptionTracker),
 		histograms:           types.NewHPointRingBuffer(memoryConsumptionTracker),
+		stepData:             &types.RangeVectorStepData{},
 	}
 }
 
@@ -74,13 +74,14 @@ func (s *Subquery) NextSeries(ctx context.Context) error {
 	return nil
 }
 
-func (s *Subquery) NextStepSamples() (types.RangeVectorStepData, error) {
+func (s *Subquery) NextStepSamples() (*types.RangeVectorStepData, error) {
 	if s.nextStepT > s.ParentQueryTimeRange.EndT {
-		return types.RangeVectorStepData{}, types.EOS
+		return nil, types.EOS
 	}
 
-	stepT := s.nextStepT
-	rangeEnd := stepT
+	s.stepData.StepT = s.nextStepT
+	rangeEnd := s.nextStepT
+	s.nextStepT += s.ParentQueryTimeRange.IntervalMilliseconds
 
 	if s.SubqueryTimestamp != nil {
 		// Timestamp from @ modifier takes precedence over query evaluation timestamp.
@@ -93,17 +94,12 @@ func (s *Subquery) NextStepSamples() (types.RangeVectorStepData, error) {
 	s.floats.DiscardPointsBefore(rangeStart)
 	s.histograms.DiscardPointsBefore(rangeStart)
 
-	s.nextStepT += s.ParentQueryTimeRange.IntervalMilliseconds
-	s.floatView = s.floats.ViewUntilSearchingForwards(rangeEnd, s.floatView)
-	s.histogramView = s.histograms.ViewUntilSearchingForwards(rangeEnd, s.histogramView)
+	s.stepData.Floats = s.floats.ViewUntilSearchingForwards(rangeEnd, s.stepData.Floats)
+	s.stepData.Histograms = s.histograms.ViewUntilSearchingForwards(rangeEnd, s.stepData.Histograms)
+	s.stepData.RangeStart = rangeStart
+	s.stepData.RangeEnd = rangeEnd
 
-	return types.RangeVectorStepData{
-		Floats:     s.floatView,
-		Histograms: s.histogramView,
-		StepT:      stepT,
-		RangeStart: rangeStart,
-		RangeEnd:   rangeEnd,
-	}, nil
+	return s.stepData, nil
 }
 
 func (s *Subquery) StepCount() int {