--- a/pkg/streamingpromql/types/data.go
+++ b/pkg/streamingpromql/types/data.go
@@ -84,25 +84,15 @@ func (i *InstantVectorSeriesDataIterator) Next() (t int64, f float64, h *histogr
 //   - RangeStart is 1712015700000 (2024-04-01T23:55:00Z)
 //   - RangeEnd is 1712016000000 (2024-04-02T00:00:00Z)
 type RangeVectorStepData struct {
-	// Floats contains the float samples for this time step, and possibly points beyond the end of the
-	// selected range. Callers should compare points' timestamps to RangeEnd.
-	//
-	// The ring buffer must not be modified, including closing it, as RangeVectorOperator implementations
-	// may return the same ring buffer for subsequent steps and reuse the same points, if the ranges for
-	// both steps overlap.
-	Floats *FPointRingBuffer
+	// Floats contains the float samples for this time step.
+	Floats FPointRingBufferView
 
-	// Histograms contains the histogram samples for this time step, and possibly points beyond the end of the
-	// selected range. Callers should compare points' timestamps to RangeEnd.
-	//
-	// The ring buffer must not be modified, including closing it, as RangeVectorOperator implementations
-	// may return the same ring buffer for subsequent steps and reuse the same points, if the ranges for
-	// both steps overlap.
+	// Histograms contains the histogram samples for this time step.
 	//
 	// FloatHistogram instances in the buffer must not be modified as they may be returned for subsequent steps.
 	// FloatHistogram instances that are retained after the next call to NextStepSamples must be copied, as they
 	// may be modified on subsequent calls to NextStepSamples.
-	Histograms *HPointRingBuffer
+	Histograms HPointRingBufferView
 
 	// StepT is the timestamp of this time step.
 	StepT int64