--- a/pkg/streamingpromql/types/fpoint_ring_buffer.go
+++ b/pkg/streamingpromql/types/fpoint_ring_buffer.go
@@ -165,7 +165,7 @@ func (v FPointRingBufferView) UnsafePoints() (head []promql.FPoint, tail []promq
 
 	if endOfHeadSegment > len(v.buffer.points) {
 		// Need to wrap around.
-		endOfTailSegment := endOfHeadSegment % len(v.buffer.points)
+		endOfTailSegment := endOfHeadSegment - len(v.buffer.points)
 		endOfHeadSegment = len(v.buffer.points)
 		return v.buffer.points[v.buffer.firstIndex:endOfHeadSegment], v.buffer.points[0:endOfTailSegment]
 	}